feat: 장바구니 계산 로직 리팩토링 및 쿠폰 기능 추가
- cartCalculations.ts 파일 삭제 및 관련 로직을 libs 폴더로 이동하여 구조 개선
- coupons.ts 및 discounts.ts 파일 추가하여 쿠폰 및 할인 관련 기능 구현
- CartPage에서 상품 가격 표시 로직을 getDisplayPrice 함수로 변경하여 가독성 향상
- 장바구니 총액 계산 및 할인 요약 정보를 위한 새로운 함수 추가
